Began  initializaiton of network
